requirements.txt
mysqlclient
Appium-Python-Client
Django==4.2
django-bootstrap5
django-environ
django-rosetta
pillow
python-gettext

